push

+++ Assets/ResetButton.cs
@@ -5,11 +5,32 @@ using UnityEngine;
 
 public class ResetButton : MonoBehaviour
 {
+    float timer = 0f;
+
     private void OnTriggerEnter(Collider other) {
         if (other.tag == "Controller") {
             Tween.Bounce(transform);
-            PlayerPrefs.DeleteAll();
-            DisplayGlobalScore.Instance.UpdateUI();
+            timer = 0f;
+        }
+    }
+
+    private void OnTriggerExit(Collider other) {
+        if (other.tag == "Controller") {
+            Tween.Bounce(transform);
+            timer = 0f;
+        }
+    }
+
+    private void OnTriggerStay(Collider other) {
+        if (other.tag == "Controller") {
+            timer += Time.deltaTime;
+            if ( timer >= 2f) {
+                Tween.Bounce(transform);
+                PlayerPrefs.DeleteAll();
+                DisplayGlobalScore.Instance.UpdateUI();
+                Tween.Bounce(transform);
+            }
+
         }
     }
 }
